+++ skrub/datasets/_utils.py
@@ -11,7 +11,9 @@ import pandas as pd
 import requests
 from sklearn.utils import Bunch
 
-DATA_HOME_ENVAR_NAME = "SKRUB_DATA_DIRECTORY"
+from .._config import get_config
+
+DATA_HOME_ENVAR_NAME = "SKB_DATA_DIRECTORY"
 DATASET_INFO = {
     "bike_sharing": {
         "urls": [
@@ -131,7 +133,8 @@ def get_data_home(data_home=None):
     user home folder.
 
     You can even customize the default data directory by setting in your environment
-    the `SKRUB_DATA_DIRECTORY` variable to an *absolute directory path*.
+    the `SKB_DATA_DIRECTORY` variable to an *absolute directory path*.
+    The deprecated `SKRUB_DATA_DIRECTORY` is still supported with a deprecation warning.
 
     Alternatively, it can be set programmatically by giving an explicit folder
     path. The '~' symbol is expanded to the user home folder.
@@ -142,7 +145,8 @@ def get_data_home(data_home=None):
     ----------
     data_home : pathlib.Path or string, optional
         The path to the skrub data directory. If `None`, the default path
-        is `~/skrub_data`.
+        is obtained from the skrub configuration (SKB_DATA_DIRECTORY environment
+        variable or `~/skrub_data`).
 
     Returns
     -------
@@ -160,12 +164,9 @@ def get_data_home(data_home=None):
         # https://docs.python.org/3/library/pathlib.html#pathlib.Path.resolve
         data_home = data_home.resolve()
     else:
-        data_home_envar = os.environ.get(DATA_HOME_ENVAR_NAME)
-
-        if data_home_envar and (path := Path(data_home_envar)).is_absolute():
-            data_home = path
-        else:
-            data_home = Path.home() / "skrub_data"
+        # Use the data folder from skrub configuration
+        config_data_folder = get_config().get("data_dir")
+        data_home = Path(config_data_folder)
 
     data_home.mkdir(parents=True, exist_ok=True)
 
@@ -197,7 +198,7 @@ def load_simple_dataset(dataset_name, data_home=None):
     """Load a dataframe and its metadata based on its dataset_name.
 
     The data will be downloaded if not found locally.
-    For e.g. the credit_fraud dataset, the filesystem will look like:
+    For example, with the credit_fraud dataset the filesystem will look like:
 
     <data_home>/
         fraud/
@@ -213,8 +214,9 @@ def load_simple_dataset(dataset_name, data_home=None):
         The name of the dataset to load. The name must be a key of `DATASET_INFO`.
 
     data_home : path, default=None
-        The directory where to download and unpack a zip file. If None, 'skrub_data'
-        is used.
+        The directory where to download and unpack a zip file. If None, the default path
+        is obtained from the skrub configuration (SKB_DATA_DIRECTORY environment
+        variable or `~/skrub_data`).
 
     Returns
     -------
@@ -226,6 +228,7 @@ def load_simple_dataset(dataset_name, data_home=None):
         - y : pd.DataFrame, target labels
         - metadata : a dictionary containing the name, description, source and target
           (description and source may be missing)
+        - path: str, the path to the full dataframe file
     """
     bunch = load_dataset_files(dataset_name, data_home)
     bunch["X"] = bunch[dataset_name]
@@ -262,12 +265,20 @@ def load_dataset_files(dataset_name, data_home):
         _extract_archive(dataset_dir, archive_path)
 
     bunch = Bunch()
-    for file_path in datafiles_dir.iterdir():
+
+    # If there is a file named <dataset_name>.csv, we load the path as the main
+    # dataset
+    path = (datafiles_dir / dataset_name).with_suffix(".csv")
+    if path.exists():
+        bunch["path"] = str(path)
+    # Loading all paths and data files
+    # Sorting to have a deterministic order for tests
+    for file_path in sorted(datafiles_dir.iterdir()):
         if file_path.suffix == ".csv":
             bunch[file_path.stem] = pd.read_csv(file_path)
         elif file_path.suffix == ".json":
             bunch[file_path.stem] = json.loads(file_path.read_text(encoding="utf-8"))
-
+        bunch[file_path.stem + "_path"] = str(file_path)
     return bunch